feat: roll driver, switch to new api.json format

+++ playwright/src/main/java/com/microsoft/playwright/DeviceDescriptor.java
@@ -23,7 +23,7 @@ public interface DeviceDescriptor {
   }
   Viewport viewport();
   String userAgent();
-  int deviceScaleFactor();
+  double deviceScaleFactor();
   boolean isMobile();
   boolean hasTouch();
   BrowserType defaultBrowserType();
